Move setUserAccountBalances to users composable
