+++ apps/web/src/composables/wallet.ts
@@ -40,7 +40,7 @@ export default function useWallet() {
   const { getLedgerAddress, loginWithLedger, sendLedgerTransaction, signLedgerMessage } = useLedger()
   const { solanaProviderList, sendSolanaTransaction, signSolanaMessage } = useSolana()
   const { getTrezorAddress, loginWithTrezor, sendTrezorTransaction, signTrezorMessage } = useTrezor()
-  const { user, getUser, getUserAnalytics, setUser, addAccount, checkIfSecondaryAddress, checkIfPrimaryUserExists, removeAccount, updatePrimaryAddress } = useUsers()
+  const { addAccount, getUser, checkIfSecondaryAddress, checkIfPrimaryUserExists, removeAccount, setUser, setUserAccountBalances, updatePrimaryAddress, user } = useUsers()
   const { getWalletConnectAddress, loginWithWalletConnect, sendWalletConnectTransaction, signWalletConnectMessage } = useWalletConnect()
 
   function getColdStorageAddress(provider: ProviderString, currency: Currency = 'ETH') {
@@ -101,8 +101,7 @@ export default function useWallet() {
           setPrimaryAddress(user?.value?.address as string) 
         }
       }
-      // TODO: Uncomment / re-enable this
-      // await setUserAccountBalances()
+      await setUserAccountBalances()
       console.log('user.value after connecting wallet :>> ', user.value)
       await refreshBreakdown()
     } catch (error: any) {
@@ -125,16 +124,6 @@ export default function useWallet() {
       return currency as Currency
     } else {
       alert('Currency selection not yet supported for this wallet provider')
-    }
-  }
-
-  async function getAccountBalance(account: Account) {
-    // TODO: Find where api endpoint is configured for ethers.
-    try {
-      const balance = await getEthersBalance(account.address)
-      return balance
-    } catch (err: any) {
-      throw new Error(err.message || 'There was an error getting the account balance')
     }
   }
 
@@ -189,7 +178,7 @@ export default function useWallet() {
     setSelectedAddress('')
     setSelectedProvider('')
     setSelectedCurrency('')
-    setUser(null)
+    setUser(undefined)
     setPrimaryAddress('')
     loadingUserWallets.value = false
     console.log('user.value :>> ', user.value)
@@ -349,25 +338,6 @@ export default function useWallet() {
 
   function setSelectedProvider (provider: ProviderString) {
     selectedProvider.value = provider
-  }
-
-  async function setUserAccountBalances() {
-    try {
-      if (user?.value?.accounts) {
-        const accounts = user.value.accounts
-        const accountsWithBalances = await Promise.all(accounts.map(async (account: Account) => {
-          const balance = await getAccountBalance(account)
-          return {
-            ...account,
-            balance
-          }
-        }))
-        user.value.accounts = accountsWithBalances
-        setUser(user.value)
-      }
-    } catch (error) {
-      throw new Error('Error setting user account balances')
-    }
   }
 
   function setWalletProviderAddresses(addresses: CryptoAddress[]) {
